refactor(shared-scripts): externalize linter hooks
