+++ .github/scripts/render-linter-report.js
@@ -1,6 +1,7 @@
 const fs = require("node:fs");
 const path = require("node:path");
 const requireEnv = require("./lib/require-env.js");
+const { loadLinterHook } = require("./lib/linter-hooks.js");
 const {
 	deriveTargetCount,
 	escapeHtml,
@@ -13,9 +14,6 @@ const {
 	readTargetKind,
 } = require("./lib/linter-shared.js");
 
-const CARGO_LINTERS = new Set(["cargo-clippy", "cargo-deny"]);
-const CARGO_DENY_CONFIG_FILE = /(?:^|\/)\.cargo\/config(?:\.toml)?$/u;
-const CARGO_DENY_POLICY_FILE = /(?:^|\/)deny\.toml$/u;
 const MAX_INLINE_PATHS = 10;
 const MAX_RENDERED_PATHS = 100;
 const DETAILS_LIMIT = 60000;
@@ -58,11 +56,12 @@ function renderReport({
 }) {
 	const linterConfig = readLinterConfig(configPath, linterName);
 	const selectedFiles = readSelectedFiles(selectedFilesPath);
-	const checkedProjects = collectProjectTargets(
+	const checkedProjects = collectProjectTargets({
+		configPath,
 		linterName,
 		sourceRepositoryPath,
 		selectedFiles,
-	);
+	});
 	const targetLines = buildTargetLines(selectedFiles, checkedProjects);
 	const result = readResult(resultPath);
 	const hasStepFailure = [selectOutcome, installOutcome, runOutcome].includes(
@@ -326,195 +325,36 @@ function formatPathSection(title, paths) {
 	];
 }
 
-function collectProjectTargets(
+function collectProjectTargets({
+	configPath,
 	linterName,
-	sourceRepositoryPath,
+	linterServicePath,
 	selectedFiles,
-) {
-	if (!CARGO_LINTERS.has(linterName)) {
+	sourceRepositoryPath,
+}) {
+	const hook = loadLinterHook({
+		configPath,
+		fileName: "render-linter-report.js",
+		linterName,
+		linterServicePath,
+	});
+
+	if (typeof hook?.collectProjectTargets !== "function") {
 		return [];
 	}
 
-	if (linterName === "cargo-deny") {
-		return collectCargoDenyProjectTargets(sourceRepositoryPath, selectedFiles);
-	}
+	const projectTargets = hook.collectProjectTargets({
+		selectedFiles,
+		sourceRepositoryPath,
+	});
 
-	const manifests = [];
-	const seen = new Set();
-
-	for (const selectedFile of selectedFiles) {
-		const manifestPath = findNearestCargoManifest(
-			sourceRepositoryPath,
-			selectedFile,
+	if (!Array.isArray(projectTargets)) {
+		throw new Error(
+			`${linterName} render-linter-report.js must return an array of project targets`,
 		);
-
-		if (manifestPath && !seen.has(manifestPath)) {
-			seen.add(manifestPath);
-			manifests.push(manifestPath);
-		}
 	}
 
-	return manifests;
-}
-
-function collectCargoDenyProjectTargets(sourceRepositoryPath, selectedFiles) {
-	const allManifests = listCargoManifests(sourceRepositoryPath);
-	const manifests = [];
-	const seen = new Set();
-
-	for (const selectedFile of selectedFiles.map((filePath) =>
-		normalizePath(filePath),
-	)) {
-		let matchingManifests = [];
-
-		if (CARGO_DENY_POLICY_FILE.test(selectedFile)) {
-			matchingManifests = allManifests.filter(
-				(manifestPath) =>
-					findCargoDenyConfig(sourceRepositoryPath, manifestPath) ===
-					selectedFile,
-			);
-		} else if (CARGO_DENY_CONFIG_FILE.test(selectedFile)) {
-			matchingManifests = allManifests.filter((manifestPath) =>
-				manifestUsesCargoConfig(manifestPath, selectedFile),
-			);
-		} else {
-			const manifestPath = findNearestCargoManifest(
-				sourceRepositoryPath,
-				selectedFile,
-			);
-
-			if (manifestPath) {
-				matchingManifests = [manifestPath];
-			}
-		}
-
-		for (const manifestPath of matchingManifests) {
-			if (!seen.has(manifestPath)) {
-				seen.add(manifestPath);
-				manifests.push(manifestPath);
-			}
-		}
-	}
-
-	return manifests;
-}
-
-function listCargoManifests(sourceRepositoryPath) {
-	const repoRoot = path.resolve(sourceRepositoryPath);
-	const manifests = [];
-	const pendingDirs = [repoRoot];
-
-	while (pendingDirs.length > 0) {
-		const currentDir = pendingDirs.pop();
-		const entries = fs.readdirSync(currentDir, { withFileTypes: true });
-
-		for (const entry of entries) {
-			if (entry.name === ".git") {
-				continue;
-			}
-
-			const entryPath = path.join(currentDir, entry.name);
-
-			if (entry.isDirectory()) {
-				pendingDirs.push(entryPath);
-				continue;
-			}
-
-			if (entry.isFile() && entry.name === "Cargo.toml") {
-				manifests.push(
-					normalizePath(path.relative(repoRoot, entryPath) || "Cargo.toml"),
-				);
-			}
-		}
-	}
-
-	return manifests.sort();
-}
-
-function findCargoDenyConfig(sourceRepositoryPath, manifestPath) {
-	const repoRoot = path.resolve(sourceRepositoryPath);
-	let currentDir = normalizePath(
-		path.posix.dirname(normalizePath(manifestPath)),
-	);
-
-	while (true) {
-		const candidate =
-			currentDir === "." ? "deny.toml" : `${currentDir}/deny.toml`;
-
-		if (fs.existsSync(path.join(repoRoot, candidate))) {
-			return candidate;
-		}
-
-		if (currentDir === ".") {
-			return null;
-		}
-
-		const parentDir = normalizePath(path.posix.dirname(currentDir));
-
-		if (parentDir === currentDir) {
-			return null;
-		}
-
-		currentDir = parentDir;
-	}
-}
-
-function manifestUsesCargoConfig(manifestPath, configPath) {
-	const manifestDir = normalizePath(
-		path.posix.dirname(normalizePath(manifestPath)),
-	);
-	const scopeDir = normalizePath(
-		path.posix.dirname(path.posix.dirname(normalizePath(configPath))),
-	);
-
-	return (
-		scopeDir === "." ||
-		manifestDir === scopeDir ||
-		manifestDir.startsWith(`${scopeDir}/`)
-	);
-}
-
-function findNearestCargoManifest(sourceRepositoryPath, filePath) {
-	const repoRoot = path.resolve(sourceRepositoryPath);
-	let currentDir = path.resolve(repoRoot, path.dirname(filePath));
-
-	if (!isWithinRoot(repoRoot, currentDir)) {
-		return null;
-	}
-
-	while (isWithinRoot(repoRoot, currentDir)) {
-		const candidate = path.join(currentDir, "Cargo.toml");
-
-		if (fs.existsSync(candidate)) {
-			return normalizePath(path.relative(repoRoot, candidate) || "Cargo.toml");
-		}
-
-		if (currentDir === repoRoot) {
-			return null;
-		}
-
-		const parentDir = path.dirname(currentDir);
-
-		if (parentDir === currentDir) {
-			return null;
-		}
-
-		currentDir = parentDir;
-	}
-
-	return null;
-}
-
-function isWithinRoot(rootPath, candidatePath) {
-	const relative = path.relative(rootPath, candidatePath);
-	return (
-		relative === "" ||
-		(!relative.startsWith("..") && !path.isAbsolute(relative))
-	);
-}
-
-function normalizePath(filePath) {
-	return filePath.replace(/\\/gu, "/");
+	return normalizeStringArray(projectTargets);
 }
 
 function buildReportSummary({
@@ -588,7 +428,6 @@ module.exports = {
 	buildReportSummary,
 	buildTargetSummary,
 	collectProjectTargets,
-	findNearestCargoManifest,
 	renderReport,
 	runFromEnv,
 };
